refactor : exception 위치 이동
